+++ configs/kth/ConvLSTM.py
@@ -15,3 +15,7 @@ filter_size = 5
 stride = 1
 patch_size = 2
 layer_norm = 0
+# training
+# lr = 5e-3
+batch_size = 16
+sched = 'onecycle'