+++ configs/kth/PhyDNet.py
@@ -2,6 +2,6 @@ method = 'PhyDNet'
 # model
 patch_size = 4
 # training
-lr = 5e-3
+# lr = 5e-3
 batch_size = 16
 sched = 'onecycle'