+++ configs/kth/SimVP.py
@@ -3,6 +3,11 @@ spatio_kernel_enc = 3
 spatio_kernel_dec = 3
 # model_type = None
 hid_S = 64
-hid_T = 512
-N_T = 8
-N_S = 4
+hid_T = 256
+N_T = 6
+N_S = 2
+# training
+lr = 1e-3
+batch_size = 16
+drop_path = 0.1
+sched = 'onecycle'