+++ configs/kth/TAU.py
@@ -12,5 +12,4 @@ alpha = 0.1
 lr = 1e-3
 batch_size = 16
 drop_path = 0.1
-batch_size = 16
 sched = 'onecycle'