+++ configs/weather/t2m_1_40625/PhyDNet.py
@@ -2,6 +2,7 @@ method = 'PhyDNet'
 # model
 patch_size = 4
 # training
-lr = 5e-3
+lr = 1e-4
 batch_size = 16
-sched = 'onecycle'
+sched = 'cosine'
+warmup_epoch = 0