+++ openstl/utils/__init__.py
@@ -10,7 +10,7 @@ from .main_utils import (set_seed, setup_multi_processes, print_log, output_name
 from .parser import create_parser
 from .predrnn_utils import (reserve_schedule_sampling_exp, schedule_sampling, reshape_patch,
                             reshape_patch_back)
-from .dmvfn_utils import LapLoss, VGGPerceptualLoss
+from .dmvfn_utils import LapLoss, MeanShift, VGGPerceptualLoss
 from .prednet_utils import get_initial_states
 from .progressbar import ProgressBar, Timer
 
@@ -24,7 +24,7 @@ __all__ = [
     'get_dataset', 'count_parameters', 'measure_throughput', 'load_config', 'update_config', 'weights_to_cpu',
     'init_dist', 'init_random_seed', 'get_dist_info', 'reduce_tensor',
     'reserve_schedule_sampling_exp', 'schedule_sampling', 'reshape_patch', 'reshape_patch_back',
-    'LapLoss', 'VGGPerceptualLoss',
+    'LapLoss', 'MeanShift', 'VGGPerceptualLoss',
     'get_initial_states',
     'ProgressBar', 'Timer', 
 ]